Seção 17 - Palestrante e Rede Social - Frontend

+++ Back/src/ProEventos.API/Helpers/Util.cs
@@ -32,9 +32,11 @@ namespace ProEventos.API.Helpers
         }
 
         public void DeleteImage(string imageName, string destino) {
-            var imagePath = Path.Combine(_env.ContentRootPath, @$"Resources/{destino}", imageName);
-            if(System.IO.File.Exists(imagePath))
-                System.IO.File.Delete(imagePath);
+            if(!string.IsNullOrEmpty(imageName)) {
+                var imagePath = Path.Combine(_env.ContentRootPath, @$"Resources/{destino}", imageName);
+                if(System.IO.File.Exists(imagePath))
+                    System.IO.File.Delete(imagePath);
+            }
         }
     }
 }
